DI for notes commands
