+++ FrEee.Core.Domain/Extensions/CommonExtensions.cs
@@ -975,7 +975,7 @@ public static class CommonExtensions
 	{
 		if (Empire.Current == null)
 			throw new InvalidOperationException("RemoveOrderClientSide is intended for client side use.");
-		var addCmd = Empire.Current.Commands.OfType<IAddOrderCommand>().SingleOrDefault(c => c.Order == order);
+		var addCmd = Empire.Current.Commands.OfType<ISetPlayerInfoCommnad>().SingleOrDefault(c => c.Order == order);
 		if (addCmd == null)
 		{
 			// not a newly added order, so create a remove command to take it off the server
