+++ FrEee.Core.Domain/Gameplay/Commands/Notes/ISetPlayerInfoCommnad.cs
@@ -9,9 +9,10 @@ using FrEee.Objects.Civilization.Orders;
 namespace FrEee.Gameplay.Commands.Orders;
 
 /// <summary>
-/// A command which adds an <see cref="IOrder"/> to an <see cref="IOrderable"/>
+/// A command which sets the <see cref="PlayerInfo"/> of an <see cref="Empire"/>.
 /// </summary>
-public interface ISetPlayerInfoCommnad
-	: IOrderCommand
+public interface ISetPlayerInfoCommand
+	: ICommand<Empire>
 {
+	PlayerInfo PlayerInfo { get; set; }
 }
