+++ FrEee.Core.Domain/Gameplay/Commands/Orders/IOrderCommandFactory.cs
@@ -13,7 +13,7 @@ namespace FrEee.Gameplay.Commands.Orders;
 /// </summary>
 public interface IOrderCommandFactory
 {
-    IAddOrderCommand AddOrder(IOrderable target, IOrder order);
+    ISetPlayerInfoCommnad AddOrder(IOrderable target, IOrder order);
 
     IOrderCommand RearrangeOrders<T>(T target, IOrder order, int deltaPosition)
         where T : IOrderable;
