+++ FrEee.Gameplay/Commands/Notes/SetPlayerNoteCommand.cs
@@ -1,0 +1,31 @@
+﻿using FrEee.Objects.Civilization;
+using FrEee.Serialization;
+using FrEee.Extensions;
+using FrEee.Objects.GameState;
+
+namespace FrEee.Gameplay.Commands.Notes;
+
+/// <summary>
+/// Sets a player note.
+/// </summary>
+public class SetPlayerNoteCommand : Command<Empire>
+{
+	public SetPlayerNoteCommand(IReferrable target, string note)
+		: base(Empire.Current)
+	{
+		Target = target;
+		Note = note;
+	}
+
+	public string Note { get; set; }
+
+	[DoNotSerialize]
+	public IReferrable Target { get { return target.Value; } set { target = value.ReferViaGalaxy(); } }
+
+	private GameReference<IReferrable> target { get; set; }
+
+	public override void Execute()
+	{
+		Executor.PlayerNotes[target] = Note;
+	}
+}
