+++ FrEee.Gameplay/Commands/Notes/SetPublicNameCommand.cs
@@ -1,0 +1,25 @@
+﻿using FrEee.Objects.Civilization;
+
+namespace FrEee.Gameplay.Commands.Notes;
+
+/// <summary>
+/// Sets the name of an object.
+/// </summary>
+public class SetPublicNameCommand : Command<INameable>
+{
+	public SetPublicNameCommand(INameable target, string name)
+		: base(target)
+	{
+		Name = name;
+	}
+
+	/// <summary>
+	/// The name to set.
+	/// </summary>
+	public string Name { get; set; }
+
+	public override void Execute()
+	{
+		Executor.Name = Name;
+	}
+}
