+++ FrEee.UI.WinForms/Controls/SpaceVehicleReport.cs
@@ -183,7 +183,7 @@ public partial class SpaceVehicleReport : UserControl, IBindable<SpaceVehicle>
 	{
 		foreach (var order in vehicle.Orders.ToArray())
 		{
-			var addCmd = Empire.Current.Commands.OfType<IAddOrderCommand>().SingleOrDefault(c => c.Order == order);
+			var addCmd = Empire.Current.Commands.OfType<ISetPlayerInfoCommnad>().SingleOrDefault(c => c.Order == order);
 			if (addCmd == null)
 			{
 				// not a newly added order, so create a remove command to take it off the server
